@@ -46,7 +46,7 @@ class AddCategory extends ConsumerWidget {
                     content: Text(
                       AppLocalization.of(context)
                           .getTranslatedVal("pls_enter_category_name"),
-                      style: Theme.of(context).textTheme.bodyText2,
+                      style: Theme.of(context).textTheme.bodyMedium,
                     ),
                     backgroundColor: Colors.redAccent,
                   ),
@@ -58,7 +58,7 @@ class AddCategory extends ConsumerWidget {
                     content: Text(
                       AppLocalization.of(context).getTranslatedVal(
                           "category_name_allowed_from_3_to_20_characters"),
-                      style: Theme.of(context).textTheme.bodyText2,
+                      style: Theme.of(context).textTheme.bodyMedium,
                     ),
                     backgroundColor: Colors.redAccent,
                   ),
@@ -73,7 +73,7 @@ class AddCategory extends ConsumerWidget {
               padding: const EdgeInsets.all(20),
               child: Text(
                 AppLocalization.of(context).getTranslatedVal("save"),
-                style: Theme.of(context).textTheme.subtitle2!.copyWith(
+                style: Theme.of(context).textTheme.titleSmall!.copyWith(
                     fontWeight: FontWeight.bold, color: Color(0xff2196F3)),
               ),
             ),
@@ -143,7 +143,7 @@ class AddCategory extends ConsumerWidget {
                 AppLocalization.of(context).getTranslatedVal("icon"),
                 style: Theme.of(context)
                     .textTheme
-                    .subtitle2!
+                    .titleSmall!
                     .copyWith(fontSize: 18, fontWeight: FontWeight.bold),
               ),
             ),
@@ -176,7 +176,7 @@ class AddCategory extends ConsumerWidget {
                 AppLocalization.of(context).getTranslatedVal("color"),
                 style: Theme.of(context)
                     .textTheme
-                    .subtitle2!
+                    .titleSmall!
                     .copyWith(fontSize: 18, fontWeight: FontWeight.bold),
               ),
             ),
